refactor: split duration with style out
