+++ composables/tasks.ts
@@ -5,6 +5,7 @@ import {
     TaskState,
     formatLocation,
 } from "./task/tokioTask";
+import type { DurationWithStyle } from "./durationWithStyle";
 import { Duration, Timestamp } from "./task/duration";
 import { fromProtoTaskStats } from "./task/tokioTaskStats";
 import { Metadata } from "~/gen/common_pb";
@@ -14,51 +15,6 @@ import {
     type Update,
 } from "~/gen/instrument_pb";
 import type { TaskDetails, TaskUpdate } from "~/gen/tasks_pb";
-
-export class DurationWithStyle {
-    value: Duration;
-    // The tailwind class to use for this duration.
-    class: string;
-
-    constructor(value: Duration, className: string) {
-        this.value = value;
-        this.class = className;
-    }
-
-    toString(): string {
-        return this.value.toString();
-    }
-
-    valueOf(): number {
-        return this.value.valueOf();
-    }
-}
-
-function getDurationWithClass(duration: Duration): DurationWithStyle {
-    const days = duration.asDays();
-    const hours = duration.asHours();
-    const minutes = duration.asMinutes();
-    const seconds = duration.asSeconds();
-    const milliseconds = duration.asMilliseconds();
-
-    let className: string;
-
-    if (days >= 1) {
-        className = "text-blue-500 dark:text-blue-300";
-    } else if (hours >= 1) {
-        className = "text-cyan-500 dark:text-cyan-300";
-    } else if (minutes >= 1) {
-        className = "text-green-500 dark:text-green-300";
-    } else if (seconds >= 1) {
-        className = "text-yellow-500 dark:text-yellow-300";
-    } else if (milliseconds >= 1) {
-        className = "text-red-500 dark:text-red-300";
-    } else {
-        className = "text-gray-500 dark:text-gray-300";
-    }
-
-    return new DurationWithStyle(duration, className);
-}
 
 function getTaskStateIconName(state: TaskState): string {
     switch (state) {
